+++ index.html
@@ -140,7 +140,7 @@
         </div>
         <div class="secc">
           <label class="d-inline mr-3"for="search">Filter By</label>
-          <select onchange="getSelectedValue(event)" class="form-select form-select-lg py-2 px-3" aria-label="Default select example">
+          <select  class="form-select form-select-lg py-2 px-3" id="formfield" aria-label="Default select example">
     
             <option selected>Preffered Name</option>
             <option value="1">Office</option>
@@ -302,7 +302,11 @@
 
 
     <!-- bootstrap 5 cdn and font awesome scripts -->
-<script src="Script/script.js"></script>
+<script type="module" src="Script/script.js"></script>
+<!-- <script  src="Script/validations.js"></script>
+<script  src="Script/search.js"></script>
+<script  src="Script/locStorage.js"></script>
+<script  src="Script/view.js"></script> -->
     <script src="https://kit.fontawesome.com/afb6be13e3.js" crossorigin="anonymous"></script>
     <script
     src="http://code.jquery.com/jquery-3.3.1.min.js"
